+++ .changeset/red-toys-vanish.md
@@ -1,0 +1,9 @@
+---
+'@modern-js/runtime': minor
+'@modern-js/plugin-tailwindcss': minor
+'@modern-js/utils': minor
+---
+
+chore(tailwindcss): merge `@modern-js/plugin-design-token` to `@modern-js/plugin-tailwindcss`
+
+chore(tailwindcss): 合并 `@modern-js/plugin-design-token` 至 `@modern-js/plugin-tailwindcss`
