+++ packages/cli/core/src/config/types/index.ts
@@ -406,6 +406,7 @@ export interface UserConfig {
    * @requires `electron` plugin
    */
   electron?: ElectronConfig;
+  // eslint-disable-next-line max-lines
 }
 
 export type ConfigParam =
