--- packages/cli/plugin-design-token/CHANGELOG.md
@@ -1,110 +1,0 @@
-# @modern-js/plugin-design-token
-
-## 1.0.11
-
-### Patch Changes
-
-- 9377d2d9d: feat: support addPlugins util in tools.postcss
-- Updated dependencies [9377d2d9d]
-- Updated dependencies [8c9ad1749]
-  - @modern-js/utils@1.7.7
-
-## 1.0.10
-
-### Patch Changes
-
-- a1198d509: feat: bump babel 7.18.0
-
-## 1.0.9
-
-### Patch Changes
-
-- 835e5d072: remove theme imports
-
-## 1.0.8
-
-### Patch Changes
-
-- ef69c680: copy App static api
-- Updated dependencies [6451a098]
-- Updated dependencies [d5a2cfd8]
-- Updated dependencies [437367c6]
-  - @modern-js/utils@1.7.6
-
-## 1.0.7
-
-### Patch Changes
-
-- d32f35134: chore: add modern/jest/eslint/ts config files to .npmignore
-- 16ca3151b: add init function
-- Updated dependencies [d32f35134]
-- Updated dependencies [6ae4a34ae]
-- Updated dependencies [b80229c79]
-- Updated dependencies [948cc4436]
-  - @modern-js/utils@1.7.3
-
-## 1.0.6
-
-### Patch Changes
-
-- 69a728375: fix: remove exports.jsnext:source after publish
-- Updated dependencies [cd7346b0d]
-- Updated dependencies [69a728375]
-  - @modern-js/utils@1.7.2
-
-## 1.0.5
-
-### Patch Changes
-
-- 895fa0ff: chore: using "workspace:\*" in devDependencies
-- Updated dependencies [2d155c4c]
-- Updated dependencies [123e432d]
-- Updated dependencies [e5a9b26d]
-- Updated dependencies [0b26b93b]
-- Updated dependencies [123e432d]
-- Updated dependencies [f9f66ef9]
-- Updated dependencies [592edabc]
-- Updated dependencies [895fa0ff]
-- Updated dependencies [3578913e]
-- Updated dependencies [1c3beab3]
-  - @modern-js/utils@1.6.0
-
-## 1.0.4
-
-### Patch Changes
-
-- 04ae5262: chore: bump @modern-js/utils to v1.4.1 in dependencies
-- 60f7d8bf: feat: add tests dir to npmignore
-- Updated dependencies [b8599d09]
-- Updated dependencies [60f7d8bf]
-- Updated dependencies [3bf4f8b0]
-  - @modern-js/utils@1.5.0
-
-## 1.0.3
-
-### Patch Changes
-
-- bebb39b6: chore: improve devDependencies and peerDependencies
-- Updated dependencies [132f7b53]
-  - @modern-js/utils@1.3.7
-
-## 1.0.2
-
-### Patch Changes
-
-- 10379961: feat: convert runtime plugin to new plugin
-- 61e3f623: feat: convert to new plugin
-- Updated dependencies [c2046f37]
-  - @modern-js/utils@1.3.6
-
-## 1.0.1
-
-### Patch Changes
-
-- 6891e4c2: add design-token plugin
-- Updated dependencies [b376c8d6]
-- Updated dependencies [e62c4efd]
-- Updated dependencies [5ed05e65]
-- Updated dependencies [e2a8233f]
-  - @modern-js/core@1.4.2
-  - @modern-js/runtime-core@1.2.3
