--- packages/cli/plugin-design-token/package.json
@@ -1,104 +1,0 @@
-{
-  "name": "@modern-js/plugin-design-token",
-  "description": "The meta-framework suite designed from scratch for frontend-focused modern web development.",
-  "homepage": "https://modernjs.dev",
-  "bugs": "https://github.com/modern-js-dev/modern.js/issues",
-  "repository": "modern-js-dev/modern.js",
-  "license": "MIT",
-  "keywords": [
-    "react",
-    "framework",
-    "modern",
-    "modern.js"
-  ],
-  "version": "1.0.11",
-  "jsnext:source": "./src/index.ts",
-  "types": "./src/index.ts",
-  "main": "./dist/js/node/index.js",
-  "module": "./dist/js/treeshaking/index.js",
-  "jsnext:modern": "./dist/js/modern/index.js",
-  "exports": {
-    ".": {
-      "node": {
-        "jsnext:source": "./src/index.ts",
-        "import": "./dist/js/modern/index.js",
-        "require": "./dist/js/node/index.js"
-      },
-      "default": "./dist/js/treeshaking/index.js"
-    },
-    "./runtime": "./dist/js/treeshaking/runtime/index.js",
-    "./cli": {
-      "jsnext:source": "./src/cli/index.ts",
-      "default": "./dist/js/node/cli/index.js"
-    }
-  },
-  "typesVersions": {
-    "*": {
-      ".": [
-        "./dist/types/runtime/index.d.ts"
-      ],
-      "cli": [
-        "./dist/types/cli/index.d.ts"
-      ]
-    }
-  },
-  "scripts": {
-    "prepare": "pnpm build",
-    "prepublishOnly": "only-allow-pnpm",
-    "new": "modern new",
-    "build": "wireit",
-    "test": "wireit"
-  },
-  "dependencies": {
-    "@babel/runtime": "^7.18.0",
-    "@modern-js/utils": "workspace:^1.8.0",
-    "tailwindcss": "^2.0.4",
-    "hoist-non-react-statics": "^3.3.2"
-  },
-  "devDependencies": {
-    "@modern-js/plugin-testing": "workspace:*",
-    "@modern-js/runtime-core": "workspace:*",
-    "@modern-js/core": "workspace:*",
-    "@scripts/build": "workspace:*",
-    "@types/jest": "^27",
-    "@types/hoist-non-react-statics": "^3.3.1",
-    "@types/node": "^14",
-    "@types/react": "^17",
-    "@types/react-dom": "^17",
-    "react": "^17.0.2",
-    "typescript": "^4",
-    "postcss": "^8.4.14"
-  },
-  "sideEffects": false,
-  "peerDependencies": {
-    "react": "^17.0.2"
-  },
-  "publishConfig": {
-    "registry": "https://registry.npmjs.org/",
-    "access": "public"
-  },
-  "wireit": {
-    "build": {
-      "command": "modern build",
-      "files": [
-        "src/**/*",
-        "tsconfig.json",
-        "package.json",
-        "./modern.config.js"
-      ],
-      "output": [
-        "dist/**/*"
-      ]
-    },
-    "test": {
-      "command": "modern test --passWithNoTests",
-      "files": [
-        "src/**/*",
-        "tsconfig.json",
-        "package.json",
-        "tests/**/*"
-      ],
-      "output": []
-    }
-  }
-}
