--- packages/cli/plugin-design-token/src/postcss-plugin/index.ts
@@ -1,18 +1,0 @@
-import type { Declaration } from 'postcss';
-
-export default ({
-  cssVarsHash = {},
-}: { cssVarsHash?: Record<string, string> } = {}) => ({
-  postcssPlugin: 'postcss-replace-css-vars',
-  Declaration(decl: Declaration) {
-    if (decl.value && typeof decl.value === 'string') {
-      decl.value = decl.value.replace(/--\S*/g, match => {
-        if (cssVarsHash[match]) {
-          return cssVarsHash[match];
-        }
-
-        return match;
-      });
-    }
-  },
-});
