--- packages/cli/plugin-design-token/tests/.eslintrc.js
@@ -1,6 +1,0 @@
-module.exports = {
-  extends: ['@modern-js'],
-  parserOptions: {
-    project: require.resolve('./tsconfig.json'),
-  },
-};
