--- packages/cli/plugin-design-token/tests/index.test.ts
@@ -1,7 +1,0 @@
-import plugin from '../src';
-
-describe('plugin-router', () => {
-  it('default', () => {
-    expect(plugin).toBeDefined();
-  });
-});
