--- packages/cli/plugin-design-token/tests/postcss.test.ts
@@ -1,19 +1,0 @@
-import postcss from 'postcss';
-import type { Result } from 'postcss';
-import getPostcssPlugin from '../src/postcss-plugin';
-
-describe('postcss', () => {
-  it('base use', () => {
-    const plugin = getPostcssPlugin({ cssVarsHash: { '--warning': 'red' } });
-    postcss([plugin])
-      .process(
-        `
-      color: --warning;
-    `,
-        { from: '', to: undefined },
-      )
-      .then((result: Result) => {
-        expect(result.css).toMatchSnapshot();
-      });
-  });
-});
