--- packages/cli/plugin-design-token/tests/tsconfig.json
@@ -1,12 +1,0 @@
-{
-  "extends": "@modern-js/tsconfig/base",
-  "compilerOptions": {
-    "declaration": false,
-    "jsx": "preserve",
-    "baseUrl": "./",
-    "isolatedModules": true,
-    "esModuleInterop": true,
-    "paths": {},
-    "types": ["node", "jest"]
-  }
-}
