--- packages/cli/plugin-design-token/tsconfig.json
@@ -1,15 +1,0 @@
-{
-  "extends": "@modern-js/tsconfig/base",
-  "compilerOptions": {
-    "declaration": false,
-    "jsx": "preserve",
-    "skipLibCheck": true,
-    "baseUrl": "./",
-    "paths": {
-      "@/*": ["./src/*"],
-      "@modern-js/plugin": ["node_modules/@modern-js/plugin"]
-    },
-    "types": ["node"]
-  },
-  "include": ["src"]
-}
