--- packages/cli/plugin-tailwind/src/index.ts
@@ -1,1 +1,0 @@
-export { default } from './cli';
