+++ packages/cli/plugin-tailwindcss/.npmignore
