+++ packages/cli/plugin-tailwindcss/jest.config.js
