+++ packages/cli/plugin-tailwindcss/modern.config.js
