+++ packages/cli/plugin-tailwindcss/package.json
@@ -14,20 +14,42 @@
   "version": "1.4.2",
   "jsnext:source": "./src/index.ts",
   "types": "./src/index.ts",
-  "main": "./dist/js/node/index.js",
-  "module": "./dist/js/treeshaking/index.js",
+  "main": "./dist/js/node/runtime/index.js",
+  "module": "./dist/js/treeshaking/runtime/index.js",
   "jsnext:modern": "./dist/js/modern/index.js",
   "exports": {
     ".": {
       "node": {
-        "jsnext:source": "./src/index.ts",
-        "require": "./dist/js/node/index.js"
+        "jsnext:source": "./src/runtime/index.ts",
+        "import": "./dist/js/modern/runtime/index.js",
+        "require": "./dist/js/node/runtime/index.js"
       },
-      "default": "./dist/js/node/index.js"
+      "default": "./dist/js/treeshaking/runtime/index.js"
+    },
+    "./runtime": {
+      "jsnext:source": "./src/runtime/index.ts",
+      "node": {
+        "import": "./dist/js/modern/runtime/index.js",
+        "require": "./dist/js/node/runtime/index.js"
+      },
+      "default": "./dist/js/treeshaking/runtime/index.js"
     },
     "./cli": {
-      "jsnext:source": "./src/cli.ts",
-      "default": "./dist/js/node/cli.js"
+      "jsnext:source": "./src/cli/index.ts",
+      "default": "./dist/js/node/cli/index.js"
+    }
+  },
+  "typesVersions": {
+    "*": {
+      ".": [
+        "./dist/types/runtime/index.d.ts"
+      ],
+      "cli": [
+        "./dist/types/cli/index.d.ts"
+      ],
+      "runtime": [
+        "./dist/types/runtime/index.d.ts"
+      ]
     }
   },
   "scripts": {
@@ -41,18 +63,23 @@
   "dependencies": {
     "@babel/runtime": "^7.18.0",
     "@modern-js/utils": "workspace:^1.8.0",
-    "@modern-js/plugin-design-token": "workspace:^1.0.11"
+    "tailwindcss": "^2.0.4",
+    "hoist-non-react-statics": "^3.3.2"
   },
   "devDependencies": {
     "@modern-js/core": "workspace:*",
+    "@modern-js/runtime-core": "workspace:*",
     "@scripts/build": "workspace:*",
     "@types/jest": "^27",
     "@types/node": "^14",
     "@types/tailwindcss": "^2.2.1",
+    "@types/hoist-non-react-statics": "^3.3.1",
+    "@types/react": "^17",
     "typescript": "^4",
     "tailwindcss": "^2.0.4",
     "jest": "^27",
     "postcss": "^8.4.14",
+    "react": "^17.0.2",
     "@scripts/jest-config": "workspace:*"
   },
   "peerDependencies": {
@@ -62,7 +89,7 @@
   "publishConfig": {
     "registry": "https://registry.npmjs.org/",
     "access": "public",
-    "types": "./dist/types/index.d.ts"
+    "types": "./dist/types/runtime/index.d.ts"
   },
   "wireit": {
     "build": {
