+++ packages/cli/plugin-tailwindcss/src/cli/index.ts
@@ -1,14 +1,15 @@
 import path from 'path';
 import {
+  lazyImport,
   PLUGIN_SCHEMAS,
   fs,
   CONFIG_CACHE_DIR,
   globby,
   nanoid,
   slash,
+  createRuntimeExportsUtils,
 } from '@modern-js/utils';
-import DesignTokenPlugin from '@modern-js/plugin-design-token/cli';
-import type { CliPlugin } from '@modern-js/core';
+import type { CliPlugin, NormalizedConfig } from '@modern-js/core';
 import { getTailwindConfig } from './tailwind';
 import { template, checkTwinMacroNotExist } from './utils';
 
@@ -26,15 +27,43 @@ export const getRandomTwConfigFileName = (internalDirectory: string) => {
 export default (): CliPlugin => ({
   name: '@modern-js/plugin-tailwindcss',
 
-  // support designSystem.supportStyledComponents
-  usePlugins: [DesignTokenPlugin()],
-
   setup: async api => {
     const { appDirectory, internalDirectory } = api.useAppContext();
     let internalTwConfigPath = '';
     // When reinstalling dependencies, most of the time the project will be restarted
     const notHaveTwinMacro = await checkTwinMacroNotExist(appDirectory);
 
+    let pluginsExportsUtils: any;
+    const designTokenModulePath = path.resolve(__dirname, '../../../../');
+
+    const resolveConfig = lazyImport('tailwindcss/resolveConfig', require);
+
+    const PLUGIN_IDENTIFIER = 'designToken';
+
+    const getDesignTokens = (userConfig: NormalizedConfig) => {
+      const {
+        source: { designSystem },
+      } = userConfig as NormalizedConfig & {
+        source: {
+          designSystem: Record<string, any>;
+        };
+      }; // TODO: Type to be filled
+
+      const tailwindcssConfig: Record<string, any> = {};
+
+      tailwindcssConfig.theme = designSystem ? { ...designSystem } : {};
+
+      // not use default design token when designToken.defaultTheme is false or theme is false
+      if (!designSystem) {
+        tailwindcssConfig.presets = [];
+      }
+
+      // when only designSystem exist, need remove supportStyledComponents
+      if (designSystem) {
+        delete tailwindcssConfig.theme.supportStyledComponents;
+      }
+      return resolveConfig(tailwindcssConfig).theme || {};
+    };
     return {
       prepare() {
         if (notHaveTwinMacro) {
@@ -56,12 +85,97 @@ export default (): CliPlugin => ({
         }
       },
 
+      modifyEntryImports({ entrypoint, imports }: any) {
+        const userConfig =
+          api.useResolvedConfigContext() as NormalizedConfig & {
+            source: {
+              designSystem?: {
+                supportStyledComponents?: boolean;
+              };
+            };
+          };
+        const {
+          source: { designSystem },
+        } = userConfig;
+        if (
+          typeof designSystem === 'object' &&
+          designSystem.supportStyledComponents
+        ) {
+          const designTokens = getDesignTokens(userConfig);
+          imports.push({
+            value: '@modern-js/runtime/plugins',
+            specifiers: [
+              {
+                imported: PLUGIN_IDENTIFIER,
+              },
+            ],
+            initialize: `
+  const designTokens = ${JSON.stringify(designTokens)};
+            `,
+          });
+        }
+
+        return {
+          entrypoint,
+          imports,
+        };
+      },
+
+      modifyEntryRuntimePlugins({ entrypoint, plugins }: any) {
+        const {
+          source: { designSystem },
+        } = api.useResolvedConfigContext() as NormalizedConfig & {
+          source: {
+            designSystem?: {
+              supportStyledComponents?: boolean;
+            };
+          };
+        };
+        let useSCThemeProvider = true;
+        if (designSystem) {
+          // when designSystem exist, designToken.styledComponents`s default value is false.
+          useSCThemeProvider = designSystem?.supportStyledComponents || false;
+        }
+
+        if (
+          typeof designSystem === 'object' &&
+          designSystem.supportStyledComponents
+        ) {
+          plugins.push({
+            name: PLUGIN_IDENTIFIER,
+            options: `{token: designTokens, useStyledComponentsThemeProvider: ${
+              useSCThemeProvider ? 'true' : 'false'
+            }, useDesignTokenContext: false}`,
+          });
+        }
+        return {
+          entrypoint,
+          plugins,
+        };
+      },
+
       validateSchema() {
         return PLUGIN_SCHEMAS['@modern-js/plugin-tailwindcss'];
       },
 
+      addRuntimeExports() {
+        pluginsExportsUtils.addExport(
+          `export { default as designToken } from '${designTokenModulePath}'`,
+        );
+      },
+
       config() {
+        const appContext = api.useAppContext();
+        pluginsExportsUtils = createRuntimeExportsUtils(
+          appContext.internalDirectory,
+          'plugins',
+        );
         return {
+          source: {
+            alias: {
+              '@modern-js/runtime/plugins': pluginsExportsUtils.getPath(),
+            },
+          },
           tools: {
             // TODO: Add interface about postcss config
             // TODO: In module project, also is called, but should not be called.
@@ -131,6 +245,10 @@ export default (): CliPlugin => ({
                 return false;
               });
             },
+            // TODO: support less、scss、css vars
+            // less: https://github.com/modern-js-dev/modern.js/pull/398/files#diff-f77c749e403fbf1fb676d5687bef3d7138a230331c849298d8afaff9b6afbc3dR166
+            // sass: https://github.com/modern-js-dev/modern.js/pull/398/files#diff-f77c749e403fbf1fb676d5687bef3d7138a230331c849298d8afaff9b6afbc3dR229
+            // postcss: https://github.com/modern-js-dev/modern.js/pull/398/files#diff-f77c749e403fbf1fb676d5687bef3d7138a230331c849298d8afaff9b6afbc3dR281
           },
         };
       },
