+++ packages/cli/plugin-tailwindcss/src/global.d.ts
