+++ packages/cli/plugin-tailwindcss/tests/fixtures/twin-macro/package.json
