+++ packages/cli/plugin-tailwindcss/tests/tsconfig.json
