+++ packages/solutions/module-tools/package.json
@@ -76,6 +76,7 @@
     "@speedy-js/speedy-types": "0.13.2-alpha.3",
     "@speedy-js/speedy-core": "0.13.2-alpha.3",
     "@speedy-js/speedy-plugin-es5": "0.13.2-alpha.3",
+    "less": "^4.1.1",
     "sass": "^1.45.0",
     "postcss": "^8.4.14",
     "postcss-import": "^14.0.2",
